% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2fireworks-package.R
\docType{package}
\name{r2fireworks-package}
\alias{r2fireworks}
\alias{r2fireworks-package}
\title{r2fireworks: Enhance Your 'Rmarkdown' and 'shiny' Apps with Dazzling Fireworks Celebrations}
\description{
Implementation of 'JQuery' \url{https://jquery.com} and 'CSS' styles to allow the display of fireworks on a document. Toolkit to easily incorporate celebratory splashes in 'Rmarkdown' and 'shiny' apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r2fireworks.obi.obianom.com/}
  \item Report bugs at \url{https://github.com/oobianom/r2fireworks/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
