% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{useFireworks}
\alias{useFireworks}
\title{Set up firework scripts and loader}
\usage{
useFireworks()
}
\value{
scripts to load fireworks and trigger to start fireworks
}
\description{
Calls to load fireworks to a page
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://r2social.obi.obianom.com}.
}

\examples{
connectButton(
link = "//rpkg.net",
visit.us = TRUE,
position = "left")
connectButton(
link = "//www.linkedin.com/in/oobianom",
linkedin = TRUE,
position = "right")
connectButton(
link = "//x.com/R2Rpkg",
x = TRUE, position = "inline")

# NOT styled
connectButton(
link = "//rpkg.net",
visit.us = TRUE,
position = "left")
connectButton(
link = "//www.linkedin.com/in/oobianom",
linkedin = TRUE,
plain = TRUE,
position = "right")
connectButton(
link = "//x.com/R2Rpkg",
x = TRUE, position = "inline")

}
